+++ src/core/mod.rs
@@ -200,7 +200,7 @@ pub struct MainState {
     pub showing_license_box: bool,
     pub showing_settings_box: bool,
     pub current_view: CurrentView,
-    pub previous_view: Option<CurrentView>,
+    pub previous_view: CurrentView,
     pub theme: GuiTheme,
     pub gui_theme_combobox_active: i32,
     pub default_font_height: i32,
@@ -222,7 +222,7 @@ impl MainState {
             showing_license_box: false,
             showing_settings_box: false,
             current_view: CurrentView::HOME,
-            previous_view: None,
+            previous_view: CurrentView::HOME,
             theme: GuiTheme::Light,
             gui_theme_combobox_active: GuiTheme::Light as i32,
             default_font_height: default_font_height,
@@ -236,7 +236,7 @@ impl MainState {
     }
 
     pub fn set_current_view(&mut self, view: CurrentView) -> () {
-        self.previous_view = Some(self.current_view);
+        self.previous_view = self.current_view;
         self.current_view = view;
     }